requirements.txt
streamlit
langchain
openai
chromadb
sentence-transformers
python-dotenv
pypdf
tiktoken
